Split notification service extension helper into separate target
