+++ KumulosSDK.xcodeproj/xcshareddata/xcschemes/KumulosSDK.xcscheme
@@ -14,20 +14,6 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "F8111E3219A95C8B0040E7D1"
-               BuildableName = "Alamofire.framework"
-               BlueprintName = "Alamofire iOS"
-               ReferencedContainer = "container:Carthage/Checkouts/Alamofire/Alamofire.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
                BlueprintIdentifier = "03DE7B691C84DEF700F789BA"
                BuildableName = "KSCrash.framework"
                BlueprintName = "KSCrash"
