+++ KumulosSdkSwiftExtension.podspec
@@ -1,6 +1,6 @@
 
 Pod::Spec.new do |s|
-  s.name = "KumulosSdkSwift"
+  s.name = "KumulosSdkSwiftExtension"
   s.version = "7.0.0"
   s.license = "MIT"
   s.summary = "Official Swift SDK for integrating Kumulos services with your mobile apps"
@@ -9,15 +9,10 @@ Pod::Spec.new do |s|
   s.source = { :git => "https://github.com/Kumulos/KumulosSdkSwift.git", :tag => "#{s.version}" }
 
   s.swift_version = "5.0"
-  s.ios.deployment_target = "9.0"
+  s.ios.deployment_target = "10.0"
 
-  s.source_files = "Sources/**/*.{h,m,swift}"
-  s.exclude_files = "Carthage", "Sources/Extension"
-  s.module_name = "KumulosSDK"
-  s.preserve_path = 'upload_dsyms.sh'
-
-  s.prepare_command = 'chmod +x upload_dsyms.sh'
-
-  s.dependency "KSCrash", "~> 1.15.19"
+  s.source_files = "Sources/Extension/**/*.{h,m,swift}"
+  s.exclude_files = "Carthage"
+  s.module_name = "KumulosSDKExtension"
 
 end
