+++ Sources/Extension/KumulosSDKExtension.h
@@ -1,0 +1,19 @@
+//
+//  KumulosSDKExtension.h
+//  KumulosSDKExtension
+//
+//  Created by Christopher Wyllie on 26/11/2019.
+//  Copyright © 2019 Kumulos. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for KumulosSDKExtension.
+FOUNDATION_EXPORT double KumulosSDKExtensionVersionNumber;
+
+//! Project version string for KumulosSDKExtension.
+FOUNDATION_EXPORT const unsigned char KumulosSDKExtensionVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <KumulosSDKExtension/PublicHeader.h>
+
+
